NODE-197, chore: rename pallet
