+++ pallets/cccp-transfer-manager/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
-name = "pallet-cccp-fast-transfer"
+name = "pallet-cccp-transfer-manager"
 version = "1.0.0"
-description = "A pallet for CCCP-V2 (Cross-Chain Communication Protocol) Fast Transfer"
+description = "A pallet for CCCP-V2 (Cross-Chain Communication Protocol) Transfer Manager"
 authors = { workspace = true }
 homepage = { workspace = true }
 edition = { workspace = true }
