+++ pallets/cccp-transfer-manager/src/lib.rs
@@ -11,7 +11,7 @@ use bp_staking::MAX_AUTHORITIES;
 use frame_support::traits::Currency;
 use parity_scale_codec::{Decode, Encode};
 use scale_info::TypeInfo;
-use sp_core::{ConstU32, RuntimeDebug, H160};
+use sp_core::{ConstU32, RuntimeDebug, H160, H256};
 use sp_runtime::BoundedVec;
 
 /// Length unbounded bytes type.
@@ -19,6 +19,15 @@ pub type UnboundedBytes = Vec<u8>;
 
 /// Asset address type.
 pub type AssetId = H160;
+
+/// Asset index hash type.
+pub type AssetIndexHash = H256;
+
+#[derive(Decode, Encode, TypeInfo, Clone, PartialEq, Eq, RuntimeDebug)]
+pub enum TransferOption {
+	Fast,
+	Standard,
+}
 
 pub type BalanceOf<T> =
 	<<T as Config>::Currency as Currency<<T as frame_system::Config>::AccountId>>::Balance;
@@ -32,11 +41,17 @@ pub struct AssetCapInfo<Balance> {
 }
 
 #[derive(Decode, Encode, TypeInfo, Clone, PartialEq, Eq, RuntimeDebug)]
-pub struct FastTransfer<AccountId, Balance> {
-	/// The amount of the fast transfer.
+pub struct TransferInfo<Balance, AccountId> {
+	/// The amount of the transfer.
 	pub amount: Balance,
-	/// The socket message of the fast transfer. (status: REQUESTED|EXECUTED)
+	/// The option of the transfer.
+	pub option: TransferOption,
+	/// The initial socket message of the transfer. (status: REQUESTED)
 	pub socket_message: UnboundedBytes,
-	/// The voters of the fast transfer.
+	/// Voters of the transfer. Voting are only required for inbound requests since the source chain are non-bifrost chains.
+	/// Socket messages originated by outbound requests are internally validated by the pallet itself. (=immediately approved)
 	pub voters: BoundedVec<AccountId, ConstU32<MAX_AUTHORITIES>>,
+	/// The voting status of the transfer.
+	/// It'll only be approved when the majority of relayers voted for the request. (for inbound requests)
+	pub is_approved: bool,
 }
